backend_clean/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
gunicorn>=22,<24
requests>=2.32,<3
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
pydantic>=2.8,<3
pytest>=8,<9
pypdf>=5,<7
python-multipart>=0.0.20,<1
geopandas>=0.14,<2
pyogrio>=0.9,<1
pyproj>=3.6,<4
shapely>=2,<3
